pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "klword"
version = "0.1.0"
description = "Word report generation library based on python-docx"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Craig Cooper" }
]
keywords = ["word", "docx", "report", "python-docx"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Operating System :: OS Independent",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "python-docx>=1.1.0",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
klword = [
  "templates/*.docx",
  "templates/*.png",
  "templates/*.jpg",
  "templates/*.jpeg",
  "templates/*.bmp",
  "templates/*.gif",
  "templates/*.webp",
  "templates/*.tif",
  "templates/*.tiff",
  "templates/*.txt"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
